+++ test/oauth2/config.test.js
@@ -1,0 +1,45 @@
+let assert = require("assert");
+
+const { printPath, setupST, startST, killAllST, cleanST } = require("../utils");
+let { ProcessState } = require("../../lib/build/processState");
+let STExpress = require("../../");
+const OAuth2Recipe = require("../../lib/build/recipe/OAuth2/recipe").default;
+let { Querier } = require("../../lib/build/querier");
+const { maxVersion } = require("../../lib/build/utils");
+
+describe(`configTest: ${printPath("[test/oauth2/config.test.js]")}`, function () {
+    beforeEach(async function () {
+        await killAllST();
+        await setupST();
+        ProcessState.getInstance().reset();
+    });
+
+    after(async function () {
+        await killAllST();
+        await cleanST();
+    });
+
+    it("Test that the recipe initializes without a config obj", async function () {
+        const connectionURI = await startST();
+        STExpress.init({
+            supertokens: {
+                connectionURI,
+            },
+            appInfo: {
+                apiDomain: "api.supertokens.io",
+                appName: "SuperTokens",
+                websiteDomain: "supertokens.io",
+            },
+            recipeList: [OAuth2Recipe.init()],
+        });
+
+        // Only run for version >= 2.9
+        let querier = Querier.getNewInstanceOrThrowError(undefined);
+        let apiVersion = await querier.getAPIVersion();
+        if (maxVersion(apiVersion, "2.8") === "2.8") {
+            return;
+        }
+
+        let OAuth2Recipe = await OAuth2Recipe.getInstanceOrThrowError();
+    });
+});